requirements.txt
amqp==5.3.1
asgiref==3.8.1
billiard==4.2.1
build==1.2.2.post1
celery==5.3.6
certifi==2025.1.31
charset-normalizer==3.4.1
click==8.1.8
click-didyoumean==0.3.1
click-plugins==1.1.1
click-repl==0.3.0
Django==5.2
django-rest-passwordreset==1.5.0
djangorestframework==3.14.0
djangorestframework_simplejwt==5.5.0
docopt==0.6.2
idna==3.10
kombu==5.5.3
packaging==25.0
pip-tools==7.4.1
pipreqs==0.4.13
prompt_toolkit==3.0.51
PyJWT==2.9.0
pyproject_hooks==1.2.0
python-dateutil==2.9.0.post0
pytz==2025.2
PyYAML==6.0.2
requests==2.31.0
setuptools==79.0.1
six==1.17.0
sqlparse==0.5.3
tzdata==2025.2
ujson==5.9.0
urllib3==2.4.0
vine==5.1.0
wcwidth==0.2.13
wheel==0.45.1
yarg==0.1.10
redis>=4.5.0
drf-spectacular>=0.27.0
social-auth-app-django>=6.0.0
